+++ xarray/core/dataarray.py
@@ -4436,6 +4436,8 @@ class DataArray(AbstractArray, DataWithCoords, DataArrayArithmetic):
         dimension(s), where the indexers are given as strings containing
         Python expressions to be evaluated against the values in the array.
 
+        The values stored in dataarrays can also be referenced in queries as 'self'.
+
         Parameters
         ----------
         queries : dict, optional
@@ -4487,17 +4489,43 @@ class DataArray(AbstractArray, DataWithCoords, DataArrayArithmetic):
         <xarray.DataArray 'a' (x: 2)>
         array([3, 4])
         Dimensions without coordinates: x
+
+        >>> da = xr.DataArray(np.arange(0, 5, 1), dims="x", name=None)
+        >>> da
+        <xarray.DataArray (x: 5)>
+        array([0, 1, 2, 3, 4])
+        Dimensions without coordinates: x
+        >>> da.query(x="self > 2")
+        <xarray.DataArray (x: 2)>
+        array([3, 4])
+        Dimensions without coordinates: x
         """
 
-        ds = self._to_dataset_whole(shallow_copy=True)
-        ds = ds.query(
+        if self.name is None:
+            # Naming unnamed dataarrays as 'self' allows querying their values still
+            name = "self"
+        else:
+            # For consistency allow named datarrays to be referred to as 'self' also
+            name = self.name
+            queries = either_dict_or_kwargs(queries, queries_kwargs, "query")
+            queries = {
+                d: (q.replace("self", name) if isinstance(q, str) else q)
+                for d, q in queries.items()
+            }
+            queries_kwargs = {}
+
+        ds = self._to_dataset_whole(name=name).query(
             queries=queries,
             parser=parser,
             engine=engine,
             missing_dims=missing_dims,
             **queries_kwargs,
         )
-        return ds[self.name]
+
+        da = ds[name]
+        if name == "self":
+            da.name = None
+        return da
 
     def curvefit(
         self,