+++ doc/whats-new.rst
@@ -44,6 +44,9 @@ New Features
   By `Thomas Hirtz <https://github.com/thomashirtz>`_.
 - allow passing a function to ``combine_attrs`` (:pull:`4896`).
   By `Justus Magin <https://github.com/keewis>`_.
+- The values stored in an dataarray can now be referenced in a call to
+  py:func:`xarray.DataArray.query` via 'self' (:issue:`5492`, :pull:`5493`).
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
 
 Breaking changes
 ~~~~~~~~~~~~~~~~
@@ -73,6 +76,9 @@ Bug fixes
 - Fix the ``repr`` of :py:class:`Variable` objects with ``display_expand_data=True``
   (:pull:`5406`)
   By `Justus Magin <https://github.com/keewis>`_.
+- Fixed :py:func:`xarray.DataArray.query` to not fail with an unnamed dataarray
+  (:issue:`5492`, :pull:`5493`).
+  By `Tom Nicholas <https://github.com/TomNicholas>`_.
 
 
 Documentation